Separate systemc montgomery
